+++ popup/js/model/options.js
@@ -8,6 +8,16 @@ import { printError } from '../helper/utils.js'
  */
 export const defaultOptions = {
   //////////////////////////////////////////
+  // GENERAL OPTIONS                      //
+  //////////////////////////////////////////
+
+  /**
+   * When enabled, the extension will log some additional information like what it indexed
+   * and performance measurements
+   */
+  debug: false,
+
+  //////////////////////////////////////////
   // SEARCH OPTIONS                       //
   //////////////////////////////////////////
 
@@ -44,6 +54,36 @@ export const defaultOptions = {
    * For precise this is always 0 (no fuzzyness)
    */
   searchFuzzyness: 0.6,
+
+  //////////////////////////////////////////
+  // COLORS AND STYLE                     //
+  //////////////////////////////////////////
+
+  /**
+   * Width of the left color marker in search results in pixels
+   */
+  colorStripeWidth: 6,
+
+  /**
+   * Color for bookmark results, expressed as CSS color
+   */
+  bookmarkColor: '#54afaf',
+  /**
+   * Color for tab results, expressed as CSS color
+   */
+  tabColor: '#b89aff',
+  /**
+   * Color for history results, expressed as CSS color
+   */
+  historyColor: '#9ece2f',
+  /**
+   * Color for search results, expressed as CSS color
+   */
+  searchColor: '#e1a535',
+  /**
+   * Color for custom search results, expressed as CSS color
+   */
+  customSearchColor: '#ce5c2f',
 
   //////////////////////////////////////////
   // SEARCH SOURCES                       //